bmtree: Refactor to untemplatize it
