+++ ffi/rust/firedancer-sys/src/ballet/bmtree.rs
@@ -1,20 +1,10 @@
 pub use crate::generated::{
-    fd_bmtree20_commit,
-    fd_bmtree20_commit_append,
-    fd_bmtree20_commit_fini,
-    fd_bmtree20_commit_init,
-    fd_bmtree20_commit_leaf_cnt,
-    fd_bmtree20_commit_t,
-    fd_bmtree20_hash_leaf,
-    fd_bmtree20_node,
-    fd_bmtree20_node_t,
-    fd_bmtree32_commit,
-    fd_bmtree32_commit_append,
-    fd_bmtree32_commit_fini,
-    fd_bmtree32_commit_init,
-    fd_bmtree32_commit_leaf_cnt,
-    fd_bmtree32_commit_t,
-    fd_bmtree32_hash_leaf,
-    fd_bmtree32_node,
-    fd_bmtree32_node_t,
+    fd_bmtree_commit_append,
+    fd_bmtree_commit_fini,
+    fd_bmtree_commit_init,
+    fd_bmtree_commit_leaf_cnt,
+    fd_bmtree_commit_t,
+    fd_bmtree_hash_leaf,
+    fd_bmtree_node,
+    fd_bmtree_node_t,
 };
